requirements.txt
flask==3.1.0
supabase==2.13.0
python-dotenv==1.1.0
psycopg2-binary==2.9.9
openpyxl==3.1.5
